fix(sidebar): collapse button and tldr and sidebar
