+++ src/components/App/AppBar/index.tsx
@@ -58,7 +58,7 @@ const Header = styled(Flex).attrs({
   left: 64px;
   right: 32px;
   transition: opacity 1s;
-  z-index: 1;
+  z-index: 99;
   padding: 20px 23px;
 `
