+++ src/components/App/SecondarySidebar/index.tsx
@@ -40,7 +40,7 @@ const Wrapper = styled(Flex)(({ theme }) => ({
   height: '100vh',
   padding: '16px 20px',
   width: '100%',
-  zIndex: 1,
+  zIndex: 30,
   display: 'flex',
   [theme.breakpoints.up('sm')]: {
     width: MENU_WIDTH,
